requirements.txt
appdirs==1.4.3
appnope==0.1.0
backcall==0.1.0
beautifulsoup4==4.8.2
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
decorator==4.4.1
distlib==0.3.0
filelock==3.0.12
idna==2.8
ipython==7.12.0
ipython-genutils==0.2.0
jedi==0.16.0
numpy==1.18.1
parso==0.6.1
pbr==5.4.4
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
prompt-toolkit==3.0.3
ptyprocess==0.6.0
Pygments==2.5.2
requests==2.22.0
selenium==3.141.0
six==1.14.0
soupsieve==1.9.5
stevedore==1.32.0
traitlets==4.3.3
urllib3==1.25.8
virtualenv==20.0.4
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
wcwidth==0.1.8
